requirements.txt
fastapi
uvicorn
fsspec
requests
aiohttp
tifffile
imagecodecs
Pillow
gcsfs
pydantic
confluent-kafka
PyYAML
google-auth
